adicionando metodos no reposiorio JPA
corrigindo erros
adicionado os metodos para o JPA
