+++ SistemaDeVendas/src/main/java/com/example/SistemaDeVendas/interfacies/IDescontoFidelidade.java
@@ -4,14 +4,15 @@ import com.example.SistemaDeVendas.entities.Cliente;
 import com.example.SistemaDeVendas.entities.DescontoFidelidade;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public interface IDescontoFidelidade {
 
     public DescontoFidelidade buscarPorId(int id);
 
-    public ArrayList<DescontoFidelidade> buscarTodos();
+    public List<DescontoFidelidade> buscarTodos();
 
-    public void salvar(int id, Cliente idCliente, float valorDesconto, String dataAplicacao);
+    public void salvar(DescontoFidelidade descontoFidelidade);
 
     public void atualizar(int id, DescontoFidelidade descontoFidelidade);
